[Security Solution][Resolver]Add backdrop to pills (#78625)
