+++ x-pack/plugins/security_solution/public/resolver/view/process_event_dot.tsx
@@ -9,7 +9,7 @@ import styled from 'styled-components';
 import { htmlIdGenerator, EuiButton, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { useSelector } from 'react-redux';
 import { FormattedMessage } from '@kbn/i18n/react';
-import { NodeSubMenu } from './submenu';
+import { NodeSubMenu } from './styles';
 import { applyMatrix3 } from '../models/vector2';
 import { Vector2, Matrix3, ResolverState } from '../types';
 import { SafeResolverEvent } from '../../../common/endpoint/types';
