+++ x-pack/plugins/security_solution/public/resolver/view/styles.tsx
@@ -5,8 +5,62 @@
  */
 
 import { EuiPanel, EuiCallOut } from '@elastic/eui';
-
 import styled from 'styled-components';
+import { NodeSubMenuComponents } from './submenu';
+
+/**
+ * See `NodeSubMenuComponents`
+ */
+export const NodeSubMenu = styled(NodeSubMenuComponents)`
+  margin: 2px 0 0 0;
+  padding: 0;
+  border: none;
+  display: flex;
+  flex-flow: column;
+
+  &.options {
+    font-size: 0.8rem;
+    display: flex;
+    flex-flow: row wrap;
+    background: transparent;
+    position: absolute;
+    top: 4.5em;
+    contain: content;
+    width: 12em;
+    z-index: 2;
+    backdrop-filter: blur(5px);
+  }
+
+  &.options .item {
+    margin: 0.25ch 0.35ch 0.35ch 0;
+    padding: 0.35em 0.5em;
+    height: fit-content;
+    width: fit-content;
+    border-radius: 2px;
+    line-height: 0.8;
+  }
+
+  &.options .item button {
+    appearance: none;
+    height: fit-content;
+    width: fit-content;
+    line-height: 0.8;
+    outline-style: none;
+    border-color: transparent;
+    box-shadow: none;
+  }
+
+  &.options .item button:focus {
+    outline-style: none;
+    border-color: transparent;
+    box-shadow: none;
+    text-decoration: underline;
+  }
+
+  &.options .item button:active {
+    transform: scale(0.95);
+  }
+`;
 
 /**
  * The top level DOM element for Resolver
